requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
fonttools==4.31.2
idna==3.3
Jinja2==3.1.1
kiwisolver==1.4.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
murmurhash==1.0.6
newsapi-python==0.2.6
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.15
tqdm==4.63.1
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
wordcloud==1.8.1
